requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
supabase
youtube-transcript-api
yt-dlp